refactor(onboard): split GPU patch rollback and finalisation into a focused module
